ignore synonyms in compound filter
Synonyms are creating havoc with the compound filter because
they can have a different length which does not match the offsets
in the token stream. Synonyms are usually not in need of
decomposition, so this hack solves the problem by simply
skipping decomposition on tokens where the term length does
not match the offset positions.