disable decompounder in ngram_search analysis
The synonym generating filters for decompounding and for
classification terms cannot be used together because they
both change the position length of a token in a way that
does not take into account changes of previously run
synonym filters.

To work around this issue, this commit disables the
photon_hyphenation_decompounder for the saerch_ngram
analyser which is the only one that uses both filters.